Introduce HistoryItemOrderingService to streamline history item sorting
